requirements.txt
 pip install 
paho-mqtt
psycopg2-binary


